refactored out buddydump and slab type helpers
