+++ pwndbg/commands/slab.py
@@ -79,7 +79,7 @@ def slab(
     active_only=False,
 ) -> None:
     if not pwndbg.aglib.kernel.has_debug_info():
-        pwndbg.aglib.kernel.symbol.load_slab_typeinfo()
+        pwndbg.aglib.kernel.slab.load_slab_typeinfo()
     if command == "list":
         slab_list(filter_)
     elif command == "info":
