+++ tests/library/qemu-system/tests/test_commands_kernel.py
@@ -88,7 +88,7 @@ def test_command_slab_info():
         assert "may only be run when debugging a Linux kernel with debug" in res
         return
     if not pwndbg.aglib.kernel.has_debug_info():
-        pwndbg.aglib.kernel.symbol.load_slab_typeinfo()
+        pwndbg.aglib.kernel.slab.load_slab_typeinfo()
     for cache in pwndbg.aglib.kernel.slab.caches():
         cache_name = cache.name
         res = gdb.execute(f"slab info -v {cache_name}", to_string=True)
@@ -107,7 +107,7 @@ def test_command_slab_contains():
         assert "may only be run when debugging a Linux kernel with debug" in res
         return
 
-    pwndbg.aglib.kernel.symbol.load_slab_typeinfo()
+    pwndbg.aglib.kernel.slab.load_slab_typeinfo()
     # retrieve a valid slab object address (first address from freelist)
     addr, slab_cache = get_slab_object_address()
